Package Spec Bugfix
In the current release, the source code is included, but it still carries the `\Code\CorePlugin` folder structure, which leads to `CorePlugin.csproj` not being able to locate its source files and failing to build. This change should hopefully address the issue in the next NuGet release of this package.